# internal/module/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Core Ledger API",
    "description": "API documentation for Core Ledger system - COA Account Management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@core-ledger.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.core-ledger.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/request-coa-accounts": {
      "get": {
        "tags": ["request-coa-accounts"],
        "summary": "Get list of COA account requests",
        "description": "Get paginated list of COA account requests with filters",
        "operationId": "getRequestCoaAccountList",
        "parameters": [
          {
            "name": "request_type",
            "in": "query",
            "description": "Request type filter",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["CREATE", "EDIT"]
            }
          },
          {
            "name": "request_status",
            "in": "query",
            "description": "Request status filter",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["PENDING", "APPROVED", "REJECTED"]
            }
          },
          {
            "name": "maker_id",
            "in": "query",
            "description": "Maker user ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "checker_id",
            "in": "query",
            "description": "Checker user ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "coa_account_id",
            "in": "query",
            "description": "COA Account ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search by account_no, name, code",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCoaAccountListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["request-coa-accounts"],
        "summary": "Create COA account request",
        "description": "Create a new COA account request (CREATE or EDIT type). Request will be created with PENDING status and requires approval.",
        "operationId": "createRequestCoaAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RequestCoaAccountCreateRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RequestCoaAccountEditRequest"
                  }
                ]
              },
              "examples": {
                "create": {
                  "summary": "Create new account request",
                  "value": {
                    "request_type": "CREATE",
                    "account_data": {
                      "code": "ASSET",
                      "account_no": "ASSET:VND.Cobo.Details",
                      "name": "Asset Account",
                      "type": "ASSET",
                      "currency": "VND",
                      "status": "ACTIVE",
                      "provider": "Cobo",
                      "description": "Asset account description"
                    }
                  }
                },
                "edit": {
                  "summary": "Edit existing account request",
                  "value": {
                    "request_type": "EDIT",
                    "account_data": {
                      "account_id": 1,
                      "account_no": "ASSET:VND.Cobo.NewDetails",
                      "name": "Updated Account Name",
                      "status": "ACTIVE",
                      "description": "Updated description"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (duplicate account or invalid data)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/request-coa-accounts/{id}": {
      "get": {
        "tags": ["request-coa-accounts"],
        "summary": "Get COA account request detail",
        "description": "Get detailed information of a specific COA account request",
        "operationId": "getRequestCoaAccountDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Request ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCoaAccountDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["request-coa-accounts"],
        "summary": "Update COA account request",
        "description": "Update a COA account request. Only allowed for PENDING requests with type CREATE or REJECTED requests.",
        "operationId": "updateRequestCoaAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Request ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RequestCoaAccountCreateRequest"
                  },
                  {
                    "$ref": "#/components/schemas/RequestCoaAccountUpdateRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/request-coa-accounts/{id}/approve": {
      "post": {
        "tags": ["request-coa-accounts"],
        "summary": "Approve COA account request",
        "description": "Approve a pending COA account request. This will create or update the COA account in the core system.",
        "operationId": "approveRequestCoaAccount",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Request ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCoaAccountApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request approved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RequestCoaAccountCreateRequest": {
        "type": "object",
        "required": ["request_type", "account_data"],
        "properties": {
          "request_type": {
            "type": "string",
            "enum": ["CREATE"],
            "example": "CREATE"
          },
          "account_data": {
            "$ref": "#/components/schemas/CoaAccountDataCreate"
          }
        }
      },
      "RequestCoaAccountEditRequest": {
        "type": "object",
        "required": ["request_type", "account_data"],
        "properties": {
          "request_type": {
            "type": "string",
            "enum": ["EDIT"],
            "example": "EDIT"
          },
          "account_data": {
            "$ref": "#/components/schemas/CoaAccountDataEdit"
          }
        }
      },
...
      "RequestCoaAccountListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginationResponse"
          },
          "error": {
            "nullable": true,
            "$ref": "#/components/schemas/ResponseError"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": true
          },
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Success"
          },
          "data": {
            "type": "string",
            "example": "Tạo request thành công. Đang chờ phê duyệt."
          },
          "system": {
            "$ref": "#/components/schemas/System"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ResponseError"
          }
        }
      },
...
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ResponseError"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}